add deletion to show purchase, remove delete file
