+++ src/components/ShowPurchase/ShowPurchase.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { Link, withRouter } from 'react-router-dom'
-import { showPurchase } from '../../api/purchases'
+import { showPurchase, deletePurchase } from '../../api/purchases'
 
 const ShowPurchase = (props) => {
   const [purchase, setPurchase] = useState(null)
@@ -26,6 +26,26 @@ const ShowPurchase = (props) => {
         })
       })
   }, [])
+
+  const handleDelete = () => {
+    deletePurchase(user, match.params.purchaseId)
+      .then(() => {
+        msgAlert({
+          heading: 'Purchase Deleted',
+          message: 'Back to the list of movies that exist',
+          variant: 'success'
+        })
+      })
+      .then(() => history.push('/movies'))
+      .catch(err => {
+        msgAlert({
+          heading: 'Deletion Failed',
+          message: 'Something went wrong: ' + err.message,
+          variant: 'danger'
+        })
+      })
+  }
+
   return (
     <div>
       {purchase ? (
@@ -33,7 +53,7 @@ const ShowPurchase = (props) => {
           <h2>{purchase.purchaseProduct}</h2>
           <h2>${purchase.productPrice}</h2>
           <button onClick={handleDelete}>Delete</button>
-          <Link to={'/purchase-update/' + purchaseId}>Update Purchase</Link>
+          <Link to={'/purchase-update/' + purchase._id}>Update Purchase</Link>
         </div>
       ) : 'Loading...'}
     </div>
